requirements.txt
certifi==2017.11.5
chardet==3.0.4
diff-match-patch==20121119
Django==1.10.6
django-crispy-forms==1.7.0
django-formtools==1.0
django-import-export==0.6.1
django-pure-pagination==0.3.0
django-reversion==2.0.12
django-simple-captcha==0.4.6
django-xadmin==0.5.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.6
jdcal==1.3
MySQL-python==1.2.5
odfpy==1.3.6
openpyxl==2.4.9
Pillow==5.0.0
PyYAML==3.12
requests==2.18.4
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
website==1.0
xadmin==0.6.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
